Back/requirements.txt
appdirs==1.4.0
appnope==0.1.0
APScheduler==3.5.3
asn1crypto==0.24.0
cached-property==1.3.0
cffi==1.12.2
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
decorator==4.0.10
defusedxml==0.4.1
Django==2.1.5
django-cors-headers==2.4.0
django-request-logging==0.6.6
djangorestframework==3.9.0
ipython==5.1.0
ipython-genutils==0.1.0
isodate==0.5.4
itypes==1.1.0
Jinja2==2.10
lockfile==0.12.2
lxml==4.2.6
MarkupSafe==1.1.0
pexpect==4.2.1
pickleshare==0.7.4
pika==0.10.0
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pycparser==2.19
Pygments==2.1.3
python-dateutil==2.7.5
pytz==2016.7
requests==2.12.2
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.4
traitlets==4.3.1
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.19.1
wcwidth==0.1.7
websockets==3.2
win-unicode-console==0.5
xmltodict==0.12.0
zeep==0.14.0
